internal: Code step and init changes required to support Thumb
In order to execute in ARM's Thumb mode, we need to set the low order
bit in PC during initialization and when we step/continue.

This requires some helpers to fixup PC based upon the current processor
mode, which may be determined from the initial state or current register
states.