Hackish support for mode switching in disassembler
The CUI's disassembly view was not updating correctly when we switch
between ARM and THUMB modes. It was foolishly written to assume that we
could keep the Capstone engine configured for the mode we initially
started in.

As a quick (disgusting) fix, we create and destroy a CS engine every
time we disassemble some code, such that the current processor mode
can be passed to the disassembler initialization.

A TODO item outlines a better plan for the future...